+++ examples/base/src/models/demo.ts
@@ -1,41 +1,41 @@
-import { ModelDefault, Dispatch } from '@uiw-admin/models'
+import { RootModel } from '@uiw-admin/models'
+import { createModel } from '@rematch/core'
 import { selectById } from '../servers/demo'
 
-const demo: ModelDefault = {
+const demo = createModel<RootModel>()({
   name: 'demo',
   state: {
     drawerVisible: false,
     tableType: '',
     queryInfo: {},
-    isView: false
+    isView: false,
   },
   reducers: {
     updateState: (state: any, payload: any) => ({
       ...state,
-      ...payload
-    })
+      ...payload,
+    }),
   },
-  effects: (dispatch: any) => ({
-    async selectById (payload: any) {
-      const dph = dispatch as Dispatch
+  effects: (dispatch) => ({
+    async selectById(payload: any) {
+      const dph = dispatch
       const data = await selectById(payload)
       if (data.code === 200) {
         dph.demo.updateState({
           drawerVisible: true,
-          queryInfo: data.data || {}
+          queryInfo: data.data || {},
         })
       }
     },
-    clean () {
-      const dph = dispatch as Dispatch
+    clean() {
+      const dph = dispatch
       dph.demo.updateState({
         drawerVisible: false,
         tableType: '',
         queryInfo: {},
-        isView: false
+        isView: false,
       })
-    }
-  })
-}
-
+    },
+  }),
+})
 export default demo