+++ examples/base/src/models/global.ts
@@ -1,26 +1,27 @@
-import { Dispatch, ModelDefault } from '@uiw-admin/models'
+import { RootModel } from '@uiw-admin/models'
+import { createModel } from '@rematch/core'
 export interface GlobalState {
-  test: string;
-  [s: string]: any;
+  test: string
+  [s: string]: any
 }
 
-const global: ModelDefault = {
+const global = createModel<RootModel>()({
   name: 'global',
   state: {
-    test: '测试全局State'
+    test: '测试全局State',
   },
   reducers: {
     updateState: (state: GlobalState, payload: GlobalState) => ({
       ...state,
-      ...payload
-    })
+      ...payload,
+    }),
   },
-  effects: (dispatch: any) => ({
-    async verify () {
-      const dph = dispatch as Dispatch
+  effects: (dispatch) => ({
+    async verify() {
+      const dph = dispatch
       dph.global.updateState({ test: '测试2' })
-    }
-  })
-}
+    },
+  }),
+})
 
 export default global