+++ CHANGELOG.md
@@ -1,11 +1,34 @@
-更新日志
-===
+# 更新日志
+
 [![](https://img.shields.io/github/release/uiwjs/uiw-admin.svg)](https://github.com/uiwjs/uiw-admin/releases)
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
+## 5.1.2(未发版)
+
+- @uiw-admin/config
+  - 🌟 新增一个默认 `define` 的 `STORAGE` 属性,本地存储使用localStorage或sessionStorage ,默认 "session",可选值  local | session
+  - 🌟 `plugins`和`loader`进行兼容处理,可以直接`require`引入,[案例](https://uiwjs.github.io/uiw-admin/#/config)
+- @uiw-admin/plugins 
+  - 🐞 fix:修复多删除引入的 `createModel` 方法
+  - 🐞 fix:修复 `RematchWebpackPlugin` 生成的 `models` 类型问题
+- @uiw-admin/user-login 
+  - 🌟 feat:新增`saveField`参数
+- @uiw-admin/components
+  - 🌟 `PromForm`：增加SelectMultiple下拉多选组件
+  
+## 5.1.2-alpha.1
+
+- @uiw-admin/plugins 
+  - 🐞 fix:修复多删除引入的 `createModel` 方法
+
+## 5.1.2-alpha.0
+
+- @uiw-admin/plugins 
+  - 🐞 fix:修复 `RematchWebpackPlugin` 生成的 `models` 类型问题
 
 ## 5.1.1
+
 - @uiw-admin/components
   - 🌟 `PromForm`：更新在线预览文档
   - 🌟 `PromForm`：增加只读模式下表单项隐藏
@@ -14,16 +37,19 @@
   - 新增 `buttons` 按钮组参数 
 
 ## 5.1.0
+
 - @uiw-admin/components
   - 🌟 `PromForm`：增加customWidgetsList prop配置自定义组件
 - @uiw-admin/plugins
   - 🌟 删除`widgets`插件
 
 ## 5.0.11
+
 - @uiw-admin/components
   - 🌟 `Upload`： 更新Upload文档
   - 🌟 `PromForm`：upload listType更改为uploadType
   - 🌟 `Table`：operateButtons & searchBtns 支持自定义render
+  
 ## 5.0.10
 
 - @uiw-admin/plugins